chore: move the cutedsl kda_decode to cula.ops
