+++ BENCHMARK_KDA_DECODE_GB200.md
@@ -1,6 +1,6 @@
 # Benchmark Results - KDA Decode
 
-> Auto-generated by `benchmarks/bench_kda_decode.py` on 2026-04-18 17:06:19.
+> Auto-generated by `benchmarks/bench_kda_decode.py` on 2026-04-18 23:38:47.
 
 > **GPU:** NVIDIA GB200  |  **CUDA:** 13.0  |  **PyTorch:** 2.11.0+cu130  |  **Python:** 3.12.3
 
@@ -8,8 +8,8 @@
 
 ## Summary
 
-- v-last speedup (FLA/cuLA): avg=1.02x, min=0.94x, max=1.11x
-- k-last speedup (FLA/cuLA): avg=1.14x, min=1.01x, max=1.31x
+- v-last speedup (FLA/cuLA): avg=1.02x, min=0.94x, max=1.16x
+- k-last speedup (FLA/cuLA): avg=1.15x, min=1.01x, max=1.31x
 - Batch sizes: [1, 4, 8, 16, 32, 64, 128, 256]
 - Q/K heads (H): [8, 32, 64]
 - V heads (HV): 128
@@ -21,14 +21,14 @@
 
 | N | cuLA v-last (ms) | cuLA k-last (ms) | FLA v-last (ms) | v-last speedup | k-last speedup |
 |---|------------------:|------------------:|----------------:|---------------:|---------------:|
-| 1 | 0.0471 | 0.0451 | 0.0517 | **1.10x** | **1.15x** |
-| 4 | 0.0548 | 0.0511 | 0.0522 | **0.95x** | **1.02x** |
-| 8 | 0.0403 | 0.0386 | 0.0445 | **1.10x** | **1.15x** |
-| 16 | 0.0709 | 0.0623 | 0.0691 | **0.98x** | **1.11x** |
-| 32 | 0.1378 | 0.1208 | 0.1300 | **0.94x** | **1.08x** |
-| 64 | 0.2629 | 0.2320 | 0.2502 | **0.95x** | **1.08x** |
-| 128 | 0.5009 | 0.4274 | 0.4892 | **0.98x** | **1.14x** |
-| 256 | 0.9800 | 0.8384 | 0.9664 | **0.99x** | **1.15x** |
+| 1 | 0.0423 | 0.0408 | 0.0490 | **1.16x** | **1.20x** |
+| 4 | 0.0533 | 0.0495 | 0.0503 | **0.94x** | **1.02x** |
+| 8 | 0.0409 | 0.0379 | 0.0415 | **1.01x** | **1.09x** |
+| 16 | 0.0705 | 0.0623 | 0.0693 | **0.98x** | **1.11x** |
+| 32 | 0.1382 | 0.1194 | 0.1313 | **0.95x** | **1.10x** |
+| 64 | 0.2630 | 0.2322 | 0.2505 | **0.95x** | **1.08x** |
+| 128 | 0.5008 | 0.4275 | 0.4892 | **0.98x** | **1.14x** |
+| 256 | 0.9801 | 0.8387 | 0.9666 | **0.99x** | **1.15x** |
 
 ### Accuracy (Output)
 
@@ -62,14 +62,14 @@
 
 | N | cuLA v-last (ms) | cuLA k-last (ms) | FLA v-last (ms) | v-last speedup | k-last speedup |
 |---|------------------:|------------------:|----------------:|---------------:|---------------:|
-| 1 | 0.0469 | 0.0454 | 0.0521 | **1.11x** | **1.15x** |
-| 4 | 0.0542 | 0.0512 | 0.0518 | **0.96x** | **1.01x** |
-| 8 | 0.0459 | 0.0415 | 0.0457 | **0.99x** | **1.10x** |
-| 16 | 0.0749 | 0.0659 | 0.0706 | **0.94x** | **1.07x** |
-| 32 | 0.1319 | 0.1130 | 0.1318 | **1.00x** | **1.17x** |
-| 64 | 0.2533 | 0.2223 | 0.2515 | **0.99x** | **1.13x** |
-| 128 | 0.4660 | 0.3943 | 0.4959 | **1.06x** | **1.26x** |
-| 256 | 0.9039 | 0.7631 | 0.9709 | **1.07x** | **1.27x** |
+| 1 | 0.0455 | 0.0441 | 0.0499 | **1.09x** | **1.13x** |
+| 4 | 0.0532 | 0.0499 | 0.0505 | **0.95x** | **1.01x** |
+| 8 | 0.0407 | 0.0363 | 0.0431 | **1.06x** | **1.19x** |
+| 16 | 0.0746 | 0.0664 | 0.0705 | **0.95x** | **1.06x** |
+| 32 | 0.1314 | 0.1135 | 0.1312 | **1.00x** | **1.16x** |
+| 64 | 0.2538 | 0.2227 | 0.2516 | **0.99x** | **1.13x** |
+| 128 | 0.4659 | 0.3944 | 0.4962 | **1.06x** | **1.26x** |
+| 256 | 0.9040 | 0.7630 | 0.9710 | **1.07x** | **1.27x** |
 
 ### Accuracy (Output)
 
@@ -103,14 +103,14 @@
 
 | N | cuLA v-last (ms) | cuLA k-last (ms) | FLA v-last (ms) | v-last speedup | k-last speedup |
 |---|------------------:|------------------:|----------------:|---------------:|---------------:|
-| 1 | 0.0468 | 0.0455 | 0.0514 | **1.10x** | **1.13x** |
-| 4 | 0.0563 | 0.0523 | 0.0535 | **0.95x** | **1.02x** |
-| 8 | 0.0436 | 0.0401 | 0.0438 | **1.00x** | **1.09x** |
-| 16 | 0.0687 | 0.0597 | 0.0706 | **1.03x** | **1.18x** |
-| 32 | 0.1317 | 0.1155 | 0.1316 | **1.00x** | **1.14x** |
-| 64 | 0.2397 | 0.2029 | 0.2519 | **1.05x** | **1.24x** |
-| 128 | 0.4581 | 0.3844 | 0.4923 | **1.07x** | **1.28x** |
-| 256 | 0.8938 | 0.7465 | 0.9786 | **1.09x** | **1.31x** |
+| 1 | 0.0453 | 0.0444 | 0.0505 | **1.11x** | **1.14x** |
+| 4 | 0.0544 | 0.0502 | 0.0518 | **0.95x** | **1.03x** |
+| 8 | 0.0401 | 0.0359 | 0.0392 | **0.98x** | **1.09x** |
+| 16 | 0.0685 | 0.0605 | 0.0699 | **1.02x** | **1.15x** |
+| 32 | 0.1324 | 0.1156 | 0.1316 | **0.99x** | **1.14x** |
+| 64 | 0.2394 | 0.2020 | 0.2520 | **1.05x** | **1.25x** |
+| 128 | 0.4578 | 0.3845 | 0.4923 | **1.08x** | **1.28x** |
+| 256 | 0.8939 | 0.7465 | 0.9784 | **1.09x** | **1.31x** |
 
 ### Accuracy (Output)
